+++ pkg/chain/types.go
@@ -2,6 +2,7 @@ package chain
 
 import (
 	"github.com/safex/gosafex/internal/crypto"
+	"github.com/safex/gosafex/internal/filestore"
 	"github.com/safex/gosafex/pkg/account"
 	"github.com/safex/gosafex/pkg/balance"
 	"github.com/safex/gosafex/pkg/key"
@@ -53,3 +54,54 @@ type Transfer struct {
 	Height  uint64
 	KImage  crypto.Key
 }
+
+//FileWallet is a wrapper for an EncryptedDB that includes wallet specific data and operations
+type FileWallet struct {
+	name              string
+	db                *filestore.EncryptedDB
+	knownOutputs      []string //REMEMBER TO INITIALIZE THIS
+	unspentOutputs    []string
+	latestBlockNumber uint64
+	latestBlockHash   string
+}
+
+//OutputInfo is a syntesis of useful information to be stored concerning an output
+type OutputInfo struct {
+	outputType    string
+	blockHash     string
+	transactionID string
+	txLocked      string
+	txType        string
+}
+
+//TransactionInfo is a syntesis of useful information to be stored concerning a transaction
+type TransactionInfo struct {
+	version         uint64
+	unlockTime      uint64
+	extra           []byte
+	blockHeight     uint64
+	blockTimestamp  uint64
+	doubleSpendSeen bool
+	inPool          bool
+	txHash          string
+}
+
+//Keys used in local filewallet, for definitions see README.md
+
+const walletInfoKey = "WalletInfo"
+const outputReferenceKey = "OutReference"
+const blockReferenceKey = "BlckReference"
+const lastBlockReferenceKey = "LSTBlckReference"
+const outputTypeReferenceKey = "OutTypeReference"
+const unspentOutputReferenceKey = "UnspentOutputReference"
+const transactionInfoReferenceKey = "TransactionInfoReference"
+
+const genericDataBucketName = "Generic"
+
+const outputKeyPrefix = "Out-"
+const outputInfoPrefix = "OutInfo-"
+const blockKeyPrefix = "Blk-"
+const transactionInfoKeyPrefix = "TxInfo-"
+const outputTypeKeyPrefix = "Typ-"
+const transactionOutputReferencePrefix = "TxOuts-"
+const blockTransactionReferencePrefix = "Txs-"
